forced update on migration/package.json

+++ db/migrations/20190327173906_create_order_table.js
@@ -3,7 +3,7 @@ exports.up = function (knex, Promise) {
   return knex.schema.createTable('orders', function (table) {
     table.increments();
     table.integer('users_id').unsigned()
-    table.integer('total');
+    table.float('total');
     table.integer('ready_time');
     table.float('taxes');
     table.foreign('users_id').references("users.id")
